+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [2.0.3] - xxxx-xx-xx
 
+## Changed
+- Using `density_prior = "uniform"` by default for Tangram (#174)
+
 ## [2.0.2] - 2025-02-21
 
 ### Added